Allow compile with MSVC and Windows headers (#49)
When the boost::pool is instantiated and the code is compiled with
MSVC compiler, it can fail if Windows specific headers are included,
adding the max and min macros.

Because some part of the header file already has preprocessor guards,
this change add same solution to call to std::numeric_limits<>::max()
to avoid any issues.

On MSVC using C++20 standard mode, the header file also fires the same
issue, only with include it on code file, without require object
instantiation.